+++ src/create.js
@@ -1,6 +1,6 @@
 import {D} from './document'
 import {Template} from './_template'
-import {NodeCo} from './_node-co'
+import {Extra} from './_extra'
 import {ListK} from './_list-k'
 import {ListS} from './_list-s'
 
@@ -14,7 +14,7 @@ var svgURI = 'http://www.w3.org/2000/svg'
  * @return {!Object} Component
  */
 export function svg(tag, options) { //eslint-disable-line no-unused-vars
-	var model = new Template(NodeCo, [[D.createElementNS, svgURI, tag]])
+	var model = new Template(Extra, [[D.createElementNS, svgURI, tag]])
 	for (var i=1; i<arguments.length; ++i) model.config(arguments[i])
 	return model
 }
@@ -27,7 +27,7 @@ export function svg(tag, options) { //eslint-disable-line no-unused-vars
  * @return {!Object} Component
  */
 export function element(tagName, options) { //eslint-disable-line no-unused-vars
-	var model = new Template(NodeCo, [[D.createElement, tagName]])
+	var model = new Template(Extra, [[D.createElement, tagName]])
 	for (var i=1; i<arguments.length; ++i) model.config(arguments[i])
 	return model
 }
@@ -40,7 +40,7 @@ export function element(tagName, options) { //eslint-disable-line no-unused-vars
  * @return {!Object} Component
  */
 export function elementNS(nsURI, tag, options) { //eslint-disable-line no-unused-vars
-	var model = new Template(NodeCo, [[D.createElementNS, nsURI, tag]])
+	var model = new Template(Extra, [[D.createElementNS, nsURI, tag]])
 	for (var i=2; i<arguments.length; ++i) model.config(arguments[i])
 	return model
 }
@@ -52,7 +52,7 @@ export function elementNS(nsURI, tag, options) { //eslint-disable-line no-unused
  * @return {!Object} Component
  */
 export function text(txt, options) { //eslint-disable-line no-unused-vars
-	var model = new Template(NodeCo, [[D.createTextNode, txt]])
+	var model = new Template(Extra, [[D.createTextNode, txt]])
 	for (var i=1; i<arguments.length; ++i) model.config(arguments[i])
 	return model
 }
@@ -67,7 +67,7 @@ export function text(txt, options) { //eslint-disable-line no-unused-vars
 export function template(node, options) { //eslint-disable-line no-unused-vars
 	if (!node.cloneNode) throw Error('invalid node')
 
-	var modl = new Template(NodeCo, [[cloneNode, node]])
+	var modl = new Template(Extra, [[cloneNode, node]])
 	for (var i=1; i<arguments.length; ++i) modl.config(arguments[i])
 	return modl
 }